FachadaLogica
