+++ nbproject/private/private.xml
@@ -3,10 +3,11 @@
     <editor-bookmarks xmlns="http://www.netbeans.org/ns/editor-bookmarks/2" lastBookmarkId="0"/>
     <open-files xmlns="http://www.netbeans.org/ns/projectui-open-files/2">
         <group>
-            <file>file:/D:/Documentos/Brenda/6toSemestre/PSI/JPA_La_Yaquesita/src/Entidades/Gastos.java</file>
-            <file>file:/D:/Documentos/Brenda/6toSemestre/PSI/JPA_La_Yaquesita/src/META-INF/persistence.xml</file>
-            <file>file:/D:/Documentos/Brenda/6toSemestre/PSI/JPA_La_Yaquesita/src/Vistas/FmPrincipal.java</file>
-            <file>file:/D:/Documentos/Brenda/6toSemestre/PSI/JPA_La_Yaquesita/src/Controladores/IngredientesJpaController.java</file>
+            <file>file:/C:/Users/Zannie/Documents/Anni/ITSON/Sexto%20Semestre/Integrador/JPA_La_Yaquesita/src/FachadaLogica/IFachadaLogica.java</file>
+            <file>file:/C:/Users/Zannie/Documents/Anni/ITSON/Sexto%20Semestre/Integrador/JPA_La_Yaquesita/src/FachadaLogica/TomarOrden.java</file>
+            <file>file:/C:/Users/Zannie/Documents/Anni/ITSON/Sexto%20Semestre/Integrador/JPA_La_Yaquesita/src/FachadaLogica/FachadaLogica.java</file>
+            <file>file:/C:/Users/Zannie/Documents/Anni/ITSON/Sexto%20Semestre/Integrador/JPA_La_Yaquesita/src/Vistas/Yaqui.java</file>
+            <file>file:/C:/Users/Zannie/Documents/Anni/ITSON/Sexto%20Semestre/Integrador/JPA_La_Yaquesita/src/Vistas/FmPrincipal.java</file>
         </group>
     </open-files>
 </project-private>
