+++ src/FachadaLogica/IFachadaLogica.java
@@ -1,0 +1,27 @@
+/*
+ * To change this license header, choose License Headers in Project Properties.
+ * To change this template file, choose Tools | Templates
+ * and open the template in the editor.
+ */
+package FachadaLogica;
+
+import Entidades.Platillo;
+import java.util.ArrayList;
+
+/**
+ *
+ * @author Hot-Dogs La Yaquesita
+ */
+public interface IFachadaLogica {
+    
+    
+    public void guardarOrden(ArrayList<Platillo> platillos, String notasOrden);
+    
+    
+    
+    
+    
+    
+    
+    
+}
