+++ src/META-INF/persistence.xml
@@ -9,7 +9,7 @@
     <class>Entidades.Orden</class>
     <class>Entidades.OrdenHasPlatillo</class>
     <properties>
-      <property name="javax.persistence.jdbc.url" value="jdbc:mysql://localhost:3306/prueba?serverTimezone=UTC"/>
+      <property name="javax.persistence.jdbc.url" value="jdbc:mysql://localhost:3306/la_yaquesita?serverTimezone=UTC"/>
       <property name="javax.persistence.jdbc.user" value="root"/>
       <property name="javax.persistence.jdbc.driver" value="com.mysql.jdbc.Driver"/>
       <property name="javax.persistence.jdbc.password" value="hotdoquitos"/>
